Updated functionalities, added new entity and its controller and components
Added Search entity, with its controller, service and repository. Updated the recommended functionality, fixed some of its bugs, added a functionality which will help to get the information to the index html, so from the javascript file it will be posible to have the data to update dynamically the graph. Also removed some unnecessary code.